pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "daily-report"
version = "0.1.0"
description = "Daily intelligence aggregation system for AI/ML research"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.42",
    "httpx>=0.27",
    "pydantic>=2.0",
    "typer>=0.15",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "feedparser>=6.0",
    "PyYAML>=6.0",
    "aiohttp>=3.9",
    "tenacity>=9.0",
    "Jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "pytest-cov>=4.0",
]

[project.scripts]
dailyreport = "src.cli:app"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
